ari-skill-paper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ari-skill-paper"
version = "0.1.0"
description = "LaTeX paper writing support MCP server for academic venues"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.0",
    "litellm>=1.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
retrieval = [
    "openreview-py>=1.0",
    "semanticscholar>=0.7",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

pythonpath = ["."]
